pyproject.toml
[tool.poetry]
name = "gh-author-inspect"
version = "0.1.0"
description = "A package for identifying comments and discussions by an author on a repository."
authors = ["Peter Baumgartner <5107405+pmbaumgartner@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "gh_author_inspect", from = "src" }]

[tool.poetry.dependencies]
python = "^3.8"
python-dateutil = "^2.8.2"
python-graphql-client = "^0.4.3"
typer = "^0.7.0"
rich = "^13.2.0"
tqdm = "^4.64.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gh-author-inspect = 'gh_author_inspect.cli:cli'
